+++ apm-sniffer/apm-sdk-plugin/mongodb-3.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/mongodb/v3/interceptor/operation/WrappedMapReduceReadOperationInterceptor.java
@@ -16,18 +16,17 @@
  *
  */
 
-package org.apache.skywalking.apm.plugin.mongodb.v4.interceptor.operation;
+package org.apache.skywalking.apm.plugin.mongodb.v3.interceptor.operation;
 
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.EnhancedInstance;
 import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.InstanceConstructorInterceptor;
-import org.apache.skywalking.apm.plugin.mongodb.v4.support.MongoNamespaceInfo;
 
-public class OperationDatabaseConstructInterceptor implements InstanceConstructorInterceptor {
+public class WrappedMapReduceReadOperationInterceptor implements InstanceConstructorInterceptor {
 
     @Override
     public void onConstruct(EnhancedInstance objInst, Object[] allArguments) {
-        String databaseName = (String) allArguments[0];
-        objInst.setSkyWalkingDynamicField(new MongoNamespaceInfo(databaseName));
+            EnhancedInstance enhancedInstance = (EnhancedInstance) allArguments[0];
+            objInst.setSkyWalkingDynamicField(enhancedInstance.getSkyWalkingDynamicField());
     }
 
 }