+++ apm-sniffer/apm-sdk-plugin/mongodb-4.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/mongodb/v4/define/readOperation/AggregateOperationInstrumentation.java
@@ -26,7 +26,7 @@ import org.apache.skywalking.apm.agent.core.plugin.interceptor.enhance.ClassInst
 import org.apache.skywalking.apm.agent.core.plugin.match.ClassMatch;
 import org.apache.skywalking.apm.agent.core.plugin.match.NameMatch;
 
-import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
+import static net.bytebuddy.matcher.ElementMatchers.any;
 
 public class AggregateOperationInstrumentation extends ClassInstanceMethodsEnhancePluginDefine {
 
@@ -53,7 +53,7 @@ public class AggregateOperationInstrumentation extends ClassInstanceMethodsEnhan
 
                     @Override
                     public ElementMatcher<MethodDescription> getConstructorMatcher() {
-                        return  takesArguments(5);
+                        return  any();
                     }
 
                     @Override