+++ apm-sniffer/apm-sdk-plugin/mongodb-4.x-plugin/src/main/java/org/apache/skywalking/apm/plugin/mongodb/v4/support/MongoOperationHelper.java
@@ -22,6 +22,7 @@ import com.mongodb.internal.bulk.DeleteRequest;
 import com.mongodb.internal.bulk.InsertRequest;
 import com.mongodb.internal.bulk.UpdateRequest;
 import com.mongodb.internal.bulk.WriteRequest;
+import com.mongodb.internal.operation.AggregateOperation;
 import com.mongodb.internal.operation.CountOperation;
 import com.mongodb.internal.operation.CreateCollectionOperation;
 import com.mongodb.internal.operation.CreateIndexesOperation;
@@ -101,6 +102,9 @@ public class MongoOperationHelper {
         } else if (obj instanceof FindAndUpdateOperation) {
             BsonDocument filter = ((FindAndUpdateOperation) obj).getFilter();
             return limitFilter(filter.toString());
+        } else if (obj instanceof AggregateOperation) {
+            List<BsonDocument> pipelines = ((AggregateOperation) obj).getPipeline();
+            return getPipelines(pipelines);
         } else if (obj instanceof MapReduceToCollectionOperation) {
             BsonDocument filter = ((MapReduceToCollectionOperation) obj).getFilter();
             return limitFilter(filter.toString());
@@ -110,6 +114,18 @@ public class MongoOperationHelper {
         } else {
             return MongoConstants.EMPTY;
         }
+    }
+
+    private static String getPipelines(List<BsonDocument> pipelines) {
+        StringBuilder params = new StringBuilder();
+        for (BsonDocument pipeline : pipelines) {
+            params.append(pipeline.toString()).append(",");
+            final int filterLengthLimit = MongoPluginConfig.Plugin.MongoDB.FILTER_LENGTH_LIMIT;
+            if (filterLengthLimit > 0 && params.length() > filterLengthLimit) {
+                return params.substring(0, filterLengthLimit) + "...";
+            }
+        }
+        return params.toString();
     }
 
     private static String getFilter(List<? extends WriteRequest> writeRequestList) {